datapunk/datapunk-cortex/tests/nlp/pipeline.py
# src/nlp/pipeline.py
from typing import Dict, Any

class NLPPipeline:
    """Basic NLP pipeline implementation"""
    
    def __init__(self):
        self.cache = {}  # Simple in-memory cache
        
    async def process(self, text: str, task: str = "sentiment") -> Dict[str, Any]:
        if not text:
            raise ValueError("Empty text")
            
        if task != "sentiment":
            raise ValueError(f"Unsupported task: {task}")
            
        # Check cache
        cache_key = f"nlp:{task}:{hash(text)}"
        if cache_key in self.cache:
            return self.cache[cache_key]
            
        # Mock sentiment analysis for now
        result = {
            "task": "sentiment",
            "text": text,
            "sentiment": "POSITIVE",
            "confidence": 0.95
        }
        
        # Cache result
        self.cache[cache_key] = result
        return result